+++ src/theme/themeContext.js
@@ -1,6 +1,6 @@
 import { createContext } from "react"
 
-const ThemeContext = createContext()
+const ThemeContext = createContext({})
 
 export default ThemeContext
